+++ releases.md
@@ -9,6 +9,6 @@ All <!-- release_count starts -->3<!-- release_count ends --> of my released pro
 * **[Solid Cockpit](https://github.com/KNowledgeOnWebScale/solid-cockpit)**: [web-app-v1.0.0](https://github.com/KNowledgeOnWebScale/solid-cockpit/tree/web-app-v1.0.0) - 2026-03-03
 <br>Web application for interacting with Solid Pod resources in an integrated, user-friendly interface.
 
-* **[Personal Website](https://github.com/ecrum19/eliascrum)**: [v0.1.0](https://github.com/ecrum19/eliascrum/tree/v0.1.0) - 2026-03-02
+* **[Personal Website](https://ecrum19.github.io/eliascrum/)**: [v0.1.0](https://github.com/ecrum19/eliascrum/tree/v0.1.0) - 2026-03-02
 <br>Source repository for the personal website.
 <!-- recent_releases ends -->